chore: rename v2 endpoints
